SensorDataViewer/Server/files/docker/picamera-thumbnail/requirements.txt
avro-python3==1.10.2
boto3==1.26.138
botocore==1.29.138
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
Pillow==9.5.0
promise==2.3
pycryptodomex==3.18.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
sinetstream==1.8.0
sinetstream-kafka==1.8.0
six==1.16.0
urllib3==1.26.16
zstandard==0.21.0
